Use real shared pointers for nebulas and planets. Cleaned further core and planet code.

+++ src/modules/Nebula.hpp
@@ -1,17 +1,17 @@
 /*
  * Stellarium
  * Copyright (C) 2002 Fabien Chereau
- * 
+ *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; either version 2
  * of the License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
@@ -58,7 +58,7 @@ public:
 	virtual QString getNameI18n(void) const {return nameI18;}
 	virtual QString getEnglishName(void) const {return englishName;}
 	virtual double getAngularSize(const StelCore *core) const {return angularSize/2;}
-	
+
 	// Methods specific to Nebula
 	void setLabelColor(const Vec3f& v) {labelColor = v;}
 	void setCircleColor(const Vec3f& v) {circleColor = v;}
@@ -66,12 +66,12 @@ public:
 	//! Get the printable nebula Type.
 	//! @return the nebula type code.
 	QString getTypeString(void) const;
-	
+
 private:
-	
+
 	//! @enum NebulaType Nebula types
-	enum NebulaType 
-	{ 
+	enum NebulaType
+	{
 		NebGx,     //!< Galaxy
 		NebOc,     //!< Open star cluster
 		NebGc,     //!< Globular star cluster, usually in the Milky Way Galaxy
@@ -82,15 +82,15 @@ private:
 		NebCn,     //!< Cluster associated with nebulosity
 		NebUnknown //!< Unknown type
 	};
-	
+
 	//! Translate nebula name using the passed translator
 	void translateName(StelTranslator& trans) {nameI18 = trans.qtranslate(englishName);}
 
 	bool readNGC(char *record);
-	
+
 	void drawLabel(const StelCore* core, const StelPainter& sPainter, float maxMagLabel);
 	void drawHints(const StelPainter& sPainter, float maxMagHints);
-    
+
 	unsigned int M_nb;              // Messier Catalog number
 	unsigned int NGC_nb;            // New General Catalog number
 	unsigned int IC_nb;             // Index Catalog number
@@ -98,7 +98,7 @@ private:
 	QString nameI18;                // Nebula name
 	float mag;                      // Apparent magnitude
 	float angularSize;              // Angular size in degree
-	Vec3f XYZ;                      // Cartesian equatorial position
+	Vec3d XYZ;                      // Cartesian equatorial position
 	Vec3d XY;                       // Store temporary 2D position
 	NebulaType nType;
